Organize tests into separate logical files
