+++ Tests/DragonService-Tests/XCTestManifests.swift
@@ -3,7 +3,16 @@ import XCTest
 #if !os(macOS)
 public func allTests() -> [XCTestCaseEntry] {
     return [
-        testCase(dragon_serviceTests.allTests),
+        testCase(ChampionTests.allTests),
+        testCase(ItemTests.allTests),
+        testCase(LanguageTests.allTests),
+        testCase(MasteryTests.allTests),
+        testCase(PerkTests.allTests),
+        testCase(ProfileIconTests.allTests),
+        testCase(RealmTests.allTests),
+        testCase(RuneTests.allTests),
+        testCase(SummonerSpellTests.allTests),
+        testCase(VersionTests.allTests)
     ]
 }
 #endif
